// core/libs/graph/src/lib/max-graph/models/theme-colors.ts
/*
 * Copyright (c) 2026 Robert Bosch Manufacturing Solutions GmbH
 *
 * See the AUTHORS file(s) distributed with this work for
 * additional information regarding authorship.
 *
 * This Source Code Form is subject to the terms of the Mozilla Public
 * License, v. 2.0. If a copy of the MPL was not distributed with this
 * file, You can obtain one at https://mozilla.org/MPL/2.0/.
 *
 * SPDX-License-Identifier: MPL-2.0
 */

export interface ThemeColors {
  aspect: string;
  property: string;
  abstractProperty: string;
  operation: string;
  event: string;
  characteristic: string;
  abstractEntity: string;
  entity: string;
  constraint: string;
  trait: string;
  unit: string;
  value: string;
  entityValue: string;
  font: string;
  border: string;
  error: string;
  primary: string;
  accent: string;
  background: string;
}